backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.1.0
Django==5.2.10
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
Faker==40.1.2
idna==3.11
pillow==12.1.0
psycopg2-binary==2.9.11
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
